Fix compiler warning

+++ src/Utilities/profiling.cpp
@@ -38,7 +38,7 @@ void ThreadProfilingManager::commitResultEnd(ProfilingResult::id id)
 		{
 			result.isFinished = true;
 			auto endTime = std::chrono::time_point_cast<std::chrono::microseconds>(clock::now()).time_since_epoch().count();
-			result.duration = endTime - result.startTime;
+			result.duration = static_cast<unsigned int>(endTime - result.startTime);
 			return;
 		}
 	}
